requirements.txt
bcrypt==4.0.1
cffi==1.15.1
click==8.1.6
colorama==0.4.6
cryptography==41.0.3
Flask==2.2.5
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
itsdangerous==2.0.0
Jinja2==3.1.2
MarkupSafe==2.1.3
Pillow==9.5.0
psycopg2==2.9.7
pycparser==2.21
PyJWT==2.8.0
python-dotenv==0.21.1
SQLAlchemy==2.0.19
typing_extensions==4.7.1
Werkzeug==2.2.3
